Added firebase login support (with email). What it will do is when the user clicks the login button, it starts an activity by firebase to log the user in via email. If the user does not have an account, one is created and the user is returned to the home screen (WE NEED TO UPDATE THE UI TO ACCOMODATE THIS, i.e. remove the register button, maybe add a toast saying that the user's account was created, etc.). When the user logs in the username is passed as an extra in the intent and the main activity is started.
In MainActivity, the buttons were launching the wrong activity (finding a tutor would launch the "look for students" activity, and vice versa. I swapped the two so the right activities are being launched.

I added firebase to the app. LET ME KNOW IF THIS CAUSES THINGS TO ACT WEIRD ON YOUR SIDE.